pyproject.toml
[project]
name = "music-explorer"
version = "0.1.0"
requires-python = ">=3.12.0"
authors = [{ name = "Mark Zwart", email = "mark.zwart@pobox.com" }]
readme = "README.md"
dependencies = [
    "amqp==5.3.1",
    "aniso8601==10.0.0",
    "attrs==25.1.0",
    "bidict==0.23.1",
    "billiard==4.2.1",
    "blinker==1.9.0",
    "celery==5.4.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "dnspython==2.7.0",
    "duckdb==1.2.0",
    "duckdb_engine==0.15.0",
    "eventlet==0.39.1",
    "Flask==3.1.0",
    "flask-restplus==0.13.0",
    "flask-socketio==5.5.1",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "honcho==2.0.0",
    "idna==3.10",
    "igraph==0.11.8",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.5",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kombu==5.4.2",
    "MarkupSafe==3.0.2",
    "numpy==2.2.2",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pandas==2.2.3",
    "polars==1.22.0",
    "prompt_toolkit==3.0.50",
    "pyaml==25.1.0",
    "pyarrow==19.0.0",
    "python-dateutil==2.9.0.post0",
    "python-engineio==4.11.2",
    "python-json-logger==3.2.1",
    "python-socketio==5.12.1",
    "python3-discogs-client==2.7.1",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "redis==5.2.1",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.22.3",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "SQLAlchemy==2.0.38",
    "texttable==1.7.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "Werkzeug==3.1.3",
    "wsproto==1.2.0"
]

[project.optional-dependencies]
dev = ["black"]

[project.scripts]
load = "app_loader.main:main"
app = "app_explorer.wsgi:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]

[tool.uv]
managed = true
